Добавил выполнение 5 раздела Подключение модулей
